Added error handler cog
